+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Support for the `formsAppId` as an optional second property in the `authService.forgotPassword()` function
+
 ### Changed
 
 - update `searchCivicPlusHCMSContentItems` to call HCMS directly